refactor: Move `PreviousIterator` into `previous.rs`
