+++ src/lib.rs
@@ -1,6 +1,7 @@
 mod constants;
 mod errors;
 mod hash;
+pub mod previous;
 pub mod ratchet;
 mod salt;
 pub mod seek;
@@ -11,5 +12,6 @@ mod serde_byte_array;
 mod test_utils;
 
 pub use errors::*;
+pub use previous::PreviousIterator;
 pub use ratchet::Ratchet;
 pub use seek::RatchetSeeker;
